Move chunk splitting code to its own file
Like the code for merging chunks, this change moves the chunk
splitting code to its own file in order to improve code structure and
maintainability.
